moved cart to icon, still working on favorites
Removed cart button and put in a cart icon that does the same thing.
Favorites icon is always shown, but no functionality yet. Need to send
every intent with a "previous" key pertaining to which Activity the
intent is sent from. New table Budget in database to store budgets
related to favorites.